Consolidate reporting into report.py and trim PR scope
Merge report_charts, report_evals, report_pool_stats, and
trajectory_stats into one report.py click group. Make
experiments/evaluation/models.py the single home of the Snowball export
URI and move the experiment-scoped ArtifactStep into launch.py, leaving
experiments/models.py untouched. Drop the endpoint placement-timeout
change, split out to #8866.
