+++ experiments/post_training/curriculum_rl/report.py
@@ -1,24 +1,43 @@
 # Copyright The Marin Authors
 # SPDX-License-Identifier: Apache-2.0
 
-"""Render report charts from W&B histories.
+"""Reporting for the curriculum-RL experiment.
 
-Produces, per model family (Qwen3-0.6B plus the Snowball 67B-A2B rounds), a grade-weighted
-headline chart, per-grade validation breakouts over steps and tokens, a
-grade-attainment chart (tokens to first reach the frontier threshold at each
-grade), curriculum weight/pass-rate trajectories per arm, an empirical
-learning-velocity curve, and a JSON summary of per-arm token budgets. Raw eval
-points are drawn faintly under an EMA-smoothed line. Runs locally (W&B access
-only); seaborn is not a workspace dependency, so run with an overlay:
+Subcommands:
 
-    uv run --with seaborn python -m experiments.post_training.curriculum_rl.report_charts --out /tmp/curriculum-charts
+- ``charts``: render report charts from W&B histories. Produces, per model
+  family (Qwen3-0.6B plus the Snowball 67B-A2B rounds), a grade-weighted
+  headline chart, per-grade validation breakouts over steps and tokens, a
+  grade-attainment chart (tokens to first reach the frontier threshold at each
+  grade), curriculum weight/pass-rate trajectories per arm, an empirical
+  learning-velocity curve, and a JSON summary of per-arm token budgets. Raw
+  eval points are drawn faintly under an EMA-smoothed line. Runs locally (W&B
+  access only); seaborn is not a workspace dependency, so run with an overlay:
+
+      uv run --with seaborn python -m experiments.post_training.curriculum_rl.report charts --out /tmp/curriculum-charts
+
+- ``evals``: print aggregated Evalchemy metrics for every arm from the
+  canonical per-run ``record.json``.
+- ``pool-stats``: print per-bin row counts and grades for one built pool
+  version.
+- ``trajectories``: aggregate retained-trajectory statistics (truncation,
+  think-token usage, answer-line compliance) for one arm.
+
+``evals``, ``pool-stats``, and ``trajectories`` read the experiment bucket, so
+run them as an Iris CPU job on the training cluster (cw-us-east-02a). The
+plotting stack is imported module-wide, so every invocation needs the
+``--with seaborn`` overlay.
 """
 
+import gzip
+import io as std_io
 import itertools
 import json
 import logging
 import math
-from collections import defaultdict
+import re
+import zipfile
+from collections import Counter, defaultdict
 from pathlib import Path
 
 import click
@@ -27,12 +46,28 @@ import matplotlib as mpl
 mpl.use("Agg")
 import matplotlib.pyplot as plt
 import pandas as pd
+import pyarrow.parquet as pq
 import seaborn as sns
 import wandb
+from marin.evaluation.records import CW_RECORDS_PREFIX, RunStatus, list_records
+from marin.training.training import temporary_storage_base_path
+from rigging.filesystem.storage_path import StoragePath, prefix_join
+
+from experiments.post_training.curriculum_rl.launch import EXPERIMENT_NAME
 
 logger = logging.getLogger(__name__)
 
 WANDB_PROJECT = "marin-community/marin-curriculum-rl"
+POOL_ROOT = "s3://marin-us-east-02a/marin/documents/curriculum-rl-pool"
+
+# Trajectory aggregation (the ``trajectories`` subcommand).
+MAX_ARCHIVES = 80
+STEP_BUCKET = 20
+# Graded answer lines and \boxed{} both live at the very end of a response.
+TAIL_CHARS = 400
+THINK_TOKENS = ("<|start_think|>", "<|end_think|>", "<think>", "</think>")
+ANSWER_LINE = re.compile(r"(####\s*\S+|Answer:\s*\S+)")
+BIN_TOKENS = ("gsm8k", "svamp", "asdiv", "math", "numina", "omni", "aime", "theoremqa", "hardmath", "rg-sum", "putnam")
 VERSION_TAG = "2026.08.31"
 # Snowball arms were relaunched at micro_train_batch_size_per_gpu=4 under a
 # bumped version; the cancelled micro=1 probe lives at plain 2026.08.31 and
@@ -466,10 +501,15 @@ def plot_velocity(samples: list[tuple[float, float]], group_size: int, path: Pat
     plt.close(fig)
 
 
-@click.command(help=__doc__)
-@click.option("--out", type=click.Path(path_type=Path), required=True)
-def main(out: Path) -> None:
+@click.group(help=__doc__)
+def main() -> None:
     logging.basicConfig(level=logging.INFO)
+
+
+@main.command()
+@click.option("--out", type=click.Path(path_type=Path), required=True)
+def charts(out: Path) -> None:
+    """Render report charts and a JSON summary from W&B histories."""
     sns.set_theme(style="whitegrid", font_scale=1.05)
     out.mkdir(parents=True, exist_ok=True)
     api = wandb.Api()
@@ -509,6 +549,129 @@ def main(out: Path) -> None:
         plot_velocity(all_velocity_samples, group_size=8, path=out / "velocity-vs-pass-rate.png")
 
     logger.info("Wrote charts and summary to %s", out)
+
+
+@main.command()
+def evals() -> None:
+    """Print aggregated Evalchemy metrics for every curriculum-RL arm."""
+    for record in list_records(CW_RECORDS_PREFIX):
+        # Model names either start with the experiment name or are
+        # owner-prefixed (e.g. "power-curriculum-rl-naive").
+        if f"{EXPERIMENT_NAME}-" not in record.model.name:
+            continue
+        if record.status is not RunStatus.SUCCEEDED:
+            print(f"== {record.model.name} {record.run_id} {record.status}")
+            continue
+        for task, metrics in record.metrics.items():
+            print(f"== {record.model.name} {task} {json.dumps(metrics)}")
+
+
+@main.command(name="pool-stats")
+@click.option("--version", required=True)
+def pool_stats(version: str) -> None:
+    """Print per-bin row counts and grades for one built pool version."""
+    for filename in ("train.parquet", "validation.parquet"):
+        path = prefix_join(prefix_join(POOL_ROOT, version), filename)
+        with StoragePath(path).open("rb") as handle:
+            table = pq.read_table(handle, columns=["extra_info"])
+        counts: Counter[tuple[int, str]] = Counter()
+        for info in table.column("extra_info").to_pylist():
+            counts[(int(info["grade"]), str(info["data_source"]))] += 1
+        print(f"== {filename}")
+        for (grade, source), n in sorted(counts.items()):
+            print(json.dumps({"grade": grade, "bin": source, "rows": n}))
+
+
+def record_stats(rec: dict, reasons: tuple[str, ...]) -> dict:
+    resp = rec.get("response", {})
+    text = resp.get("text") or ""
+    extras = json.dumps(rec.get("trajectory", {}).get("environment_extras"))
+    bin_token = next((t for t in BIN_TOKENS if t in extras), "other")
+    return {
+        "step": rec.get("global_step") or 0,
+        "bin": bin_token,
+        # Failures and truncations are retained mandatorily, so the
+        # non-mandatory remainder is a hash sample of successful terminating
+        # rollouts only. None when the record id is missing from the ledger.
+        "sampled": None if not reasons else "mandatory" not in reasons,
+        "truncated": resp.get("stop_reason") == "length",
+        "think": any(t in text for t in THINK_TOKENS),
+        "answer_line": bool(ANSWER_LINE.search(text[-TAIL_CHARS:])),
+        "boxed": "\\boxed{" in text[-TAIL_CHARS:],
+        "chars": len(text),
+    }
+
+
+@main.command()
+@click.argument("arm")
+@click.argument("version")
+def trajectories(arm: str, version: str) -> None:
+    """Aggregate retained-trajectory statistics for one arm.
+
+    Reports, per 20-step bucket and per bin: truncation rate
+    (stop_reason=length), canonical think-token usage, graded answer-line
+    compliance, and \\boxed{} usage.
+    """
+    out = f"s3://marin-us-east-02a/marin/users/power/checkpoints/curriculum-rl/{arm}/{version}"
+    root = temporary_storage_base_path(out, ttl_days=14, category="skyrl")
+    traj_root = prefix_join(prefix_join(root, "attempts"), "trajectories")
+    ledger = json.loads(StoragePath(prefix_join(traj_root, "_retention_ledger.json")).read_text())
+    reasons_by_id = {rid: tuple(entry.get("reasons", ())) for rid, entry in ledger.get("records", {}).items()}
+    archives = sorted(ledger.get("archives", {}))
+    if len(archives) > MAX_ARCHIVES:
+        stride = len(archives) / MAX_ARCHIVES
+        sampled = [archives[int(i * stride)] for i in range(MAX_ARCHIVES)]
+        print(f"sampling {MAX_ARCHIVES} of {len(archives)} archives (even stride)")
+    else:
+        sampled = archives
+        print(f"reading all {len(archives)} archives")
+
+    rows = []
+    for apath in sampled:
+        payload = StoragePath(prefix_join(traj_root, apath)).read_bytes()
+        try:
+            zf = zipfile.ZipFile(std_io.BytesIO(payload))
+        except zipfile.BadZipFile as exc:
+            print("archive", apath, "unreadable:", exc)
+            continue
+        for name in zf.namelist():
+            if not name.endswith(".json.gz"):
+                continue
+            record_id = name.rsplit("/", 1)[-1].removesuffix(".json.gz")
+            reasons = reasons_by_id.get(record_id, ())
+            rows.append(record_stats(json.loads(gzip.decompress(zf.read(name))), reasons))
+    matched = sum(1 for r in rows if r["sampled"] is not None)
+    print(f"records: {len(rows)} (ledger reason matched for {matched})")
+
+    def bucket(rows_subset: list[dict], label: str) -> None:
+        n = len(rows_subset)
+        if not n:
+            return
+        trunc = sum(r["truncated"] for r in rows_subset) / n
+        think = sum(r["think"] for r in rows_subset) / n
+        ans = sum(r["answer_line"] for r in rows_subset) / n
+        boxed = sum(r["boxed"] for r in rows_subset) / n
+        chars = sum(r["chars"] for r in rows_subset) / n
+        print(
+            f"{label:>16} n={n:<6} trunc={trunc:.3f} think={think:.3f} "
+            f"answer_line={ans:.3f} boxed={boxed:.3f} avg_chars={chars:,.0f}"
+        )
+
+    for label, subset in (
+        ("success stream (hash-sampled)", [r for r in rows if r["sampled"] is True]),
+        ("failure/trunc stream (mandatory)", [r for r in rows if r["sampled"] is False]),
+    ):
+        print(f"\n==== {label} ====")
+        by_step = defaultdict(list)
+        for r in subset:
+            by_step[r["step"] // STEP_BUCKET * STEP_BUCKET].append(r)
+        for k in sorted(by_step):
+            bucket(by_step[k], f"steps {k}-{k + STEP_BUCKET - 1}")
+        by_bin = defaultdict(list)
+        for r in subset:
+            by_bin[r["bin"]].append(r)
+        for k in sorted(by_bin):
+            bucket(by_bin[k], k)
 
 
 if __name__ == "__main__":
